requirements.txt
Django==4.2.8
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
Pillow==10.1.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
whitenoise==6.6.0
celery==5.3.4
redis==5.0.1
requests==2.31.0
markdown==3.5.1
drf-yasg==1.21.7
daphne==4.0.0
django-storages==1.14.2
